+++ src/main/java/de/geheimagentnr1/manyideas_core/elements/items/ModItemsRegisterFactory.java
@@ -60,6 +60,9 @@ public class ModItemsRegisterFactory extends ItemsRegisterFactory {
 	@ObjectHolder( registryName = RegistryKeys.ITEMS, value = ManyIdeasCore.MODID + ":" + PanelJungle.registry_name )
 	public static PanelJungle PANEL_JUNGLE;
 	
+	@ObjectHolder( registryName = RegistryKeys.ITEMS, value = ManyIdeasCore.MODID + ":" + PanelMangrove.registry_name )
+	public static PanelMangrove PANEL_MANGROVE;
+	
 	@ObjectHolder( registryName = RegistryKeys.ITEMS, value = ManyIdeasCore.MODID + ":" + PanelOak.registry_name )
 	public static PanelOak PANEL_OAK;
 	
@@ -122,6 +125,7 @@ public class ModItemsRegisterFactory extends ItemsRegisterFactory {
 			RegistryEntry.create( PanelCrimson.registry_name, new PanelCrimson() ),//FINRT
 			RegistryEntry.create( PanelDarkOak.registry_name, new PanelDarkOak() ),//FINRT
 			RegistryEntry.create( PanelJungle.registry_name, new PanelJungle() ),//FINRT
+			RegistryEntry.create( PanelMangrove.registry_name, new PanelMangrove() ),//FINRT
 			RegistryEntry.create( PanelOak.registry_name, new PanelOak() ),//FINRT
 			RegistryEntry.create( PanelSpruce.registry_name, new PanelSpruce() ),//FINRT
 			RegistryEntry.create( PanelWarped.registry_name, new PanelWarped() ),//FINRT